prompt-evaluator/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-evaluator"
version = "1.0.0"
description = "AI Prompt Evaluator — A Mini RLHF Tool for comparing AI responses"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
prompt-evaluator = "evaluator.cli:main"

[tool.setuptools.packages.find]
include = ["evaluator*"]
